refactor: split plotting and report utilities
